Move transcription length stats to separate file
